Reshaped for unit tests and deployment
