+++ Hook/app.py
@@ -1,9 +1,10 @@
+import os
+
 from flask import Flask
 from flask.ext.mongoengine import MongoEngine 
 from flask.ext.github import GitHub
 from flask.ext.login import LoginManager
 from flask_environments import Environments
-import os
 
 app = Flask(
     __name__,
@@ -18,10 +19,10 @@ db = MongoEngine(app)
 github_api = GitHub(app)
 login_manager = LoginManager(app)
 
-from routes import ui
-from routes import github
-from routes import user
-from routes.api import user as users
-from routes.api import badges
-from routes.api import test
-import ui.templating
+from Hook.routes import ui
+from Hook.routes import github
+from Hook.routes import user
+from Hook.routes.api import user as users
+from Hook.routes.api import badges
+from Hook.routes.api import test
+import Hook.ui.templating
