+++ run.py
@@ -1,0 +1,3 @@
+import Hook.app
+
+Hook.app.app.run()
