+++ test/test.py
@@ -1,322 +1,192 @@
-from lxml import etree
-import MyCapytain.resources.texts.local
-import MyCapytain.resources.inventory
-import MyCapytain.common.reference
-import os
+# -*- coding: utf-8 -*-
 
-import jingtrang
-import pkg_resources
+from collections import defaultdict
+from sys import argv, stdout, exit
 import subprocess
+import time
+import concurrent.futures
 
-curr_dir = os.path.dirname(__file__)
+import utils
+import units
+import json
+import statistics
 
-EPIDOC = os.path.join(curr_dir, "../data/external/tei-epidoc.rng")
-TEI_ALL = os.path.join(curr_dir, "../data/external/tei_all.rng")
-JING = pkg_resources.resource_filename("jingtrang", "jing.jar")
+PNRT_SETTINGS = {
+    "directory" : "",
+    "reponame" : ""
+}
 
-ns = {"tei" : "http://www.tei-c.org/ns/1.0", "ti":"http://chs.harvard.edu/xmlns/cts"}
-class TESTUnit(object):
-    def __init__(self, path):
-        self.path = path
-        self.xml = None
-        self.testable = True
-        self.__logs = []
-        self.__archives = []
-        self.Text = False
-        self.urn = None
+def cover(test):
+    """ Given a dictionary, compute the coverage of one item 
 
-    @property
-    def logs(self):
-        return self.__logs
-    
-    def log(self, message):
-        self.__logs.append(">>>>>> "+ message )
-    
-    def error(self, error):
-        self.__logs.append(">>>>>> "+ str(type(error)) + " : " + str(error) )
+    :param test: Dictionary where keys represents test done on a file and value a boolean indicating passing status 
+    :type test: boolean
 
-    def flush(self):
-        self.__archives = self.__archives + self.__logs
-        self.__logs = []
+    :returns: Passing status 
+    :rtype: dict
 
-    def parsable(self):
-        """ Check and parse the xml file
-
-        :returns: Indicator of success and messages
-        :rtype: boolean
-
-        """
-        try:
-            f = open(self.path)
-            xml = etree.parse(f)
-            self.xml = xml
-            self.testable = True
-            self.log("Parsed")
-            f.close()
-        except Exception as e:
-            self.testable = False
-            self.error(e)
-        finally:
-            yield self.testable
-
-
-class INVUnit(TESTUnit):
-    """ CTS testing object
-
-    :param path: Path to the file
-    :type path: basestring
-    """
-
-    tests = ["parsable", "capitain", "metadata", "check_urns"]
-    readable = {
-        "parsable" : "File parsing",
-        "capitain" : "MyCapytain parsing",
-        "metadata" : "Metadata availability",
-        "check_urns" : "URNs testing"
+    """ 
+    results = list(test.values())
+    return {
+        "units" : test,
+        "coverage" : len([v for v in results if v is True])/len(results)*100,
+        "status" : False not in results
     }
 
-    def capitain(self):
-        """ Load the file in MyCapytain
-        """
-        textgroup = "textgroup" in self.xml.getroot().tag
-        work = not textgroup and "work" in self.xml.getroot().tag
-        if textgroup:
-            self.type = "textgroup"
-            self.log("TextGroup detected")
-            self.Text = MyCapytain.resources.inventory.TextGroup(resource=self.xml.getroot())
-        elif work:
-            self.type = "work"
-            self.log("Work detected")
-            self.Text = MyCapytain.resources.inventory.Work(resource=self.xml.getroot())
-        else:
-            self.log("Nothing detected")
+def prnt(data):
+    """ Print data and flush the stdout to be able to retrieve information line by line on another tool
 
-        if self.Text:
-            yield True
-        else:
-            yield False
+    :param data: Data to be printed
+    :type data: str
 
-    def metadata(self):
-        if self.type == "textgroup":
+    """
+    print(data.replace(PNRT_SETTINGS["directory"], PNRT_SETTINGS["reponame"]), flush=True)
 
-            groups = len(self.Text.metadata["groupname"].children)
-            self.log("{0} groupname found".format(str(groups)))
-            yield groups > 0
 
-        elif self.type == "work":
+def do_test(f, tei, epidoc, verbose, inventory, results, passing):
+    """ Do test for a file and print the results
 
-            titles = len(self.Text.metadata["title"].children)
-            self.log("{0} titles found".format(titles))
-            status = titles > 0
+    :param f: Path of the file to be tested
+    :type f: str
 
-            texts = len(self.Text.texts)
-            labels = len(
-                [
-                    text for text in self.Text.texts.values()
-                    if len(text.metadata["label"].children) > 0
-                ]
-            )
+    :returns: List of status informations
+    :rtype: list
+    """
+    logs = []
+    if not inventory:
+        inventory = []
+    if f.endswith("__cts__.xml"):
+        t = units.INVUnit(f)
+        logs.append(">>>> Testing "+ f)
 
-            self.log("{0}/{1} file(s) with labels".format(labels, texts))
-            status = status and labels == texts
-
-            descs = len(
-                [
-                    text for text in self.Text.texts.values()
-                    if len(text.metadata["description"].children) > 0
-                ]
-            )
-            self.log("{0}/{1} file(s) with descs".format(descs, texts))
-            status = status and labels == descs
-
-            yield status
-        else:
-            yield False
-
-    def check_urns(self):
-        if self.type == "textgroup":
-            urns = [
-                urn
-                for urn in self.xml.xpath("//ti:textgroup/@urn", namespaces=ns)
-                if len(MyCapytain.common.reference.URN(urn)) == 3
-            ]
-            self.log("Group urn :" + "".join(self.xml.xpath("//ti:textgroup/@urn", namespaces=ns)))
-            yield len(urns) == 1
-        elif self.type == "work":
-            worksUrns = [
-                    urn
-                    for urn in self.xml.xpath("//ti:work/@urn", namespaces=ns)
-                    if len(MyCapytain.common.reference.URN(urn)) == 4
-                ] + [
-                    urn
-                    for urn in self.xml.xpath("//ti:work/@groupUrn", namespaces=ns)
-                    if len(MyCapytain.common.reference.URN(urn)) == 3
-                ]
-            self.log("Group urn : " + "".join(self.xml.xpath("//ti:work/@groupUrn", namespaces=ns)))
-            self.log("Work urn : " + "".join(self.xml.xpath("//ti:work/@urn", namespaces=ns)))
-
-            texts = self.xml.xpath("//ti:edition|//ti:translation", namespaces=ns)
-            workUrnsText = []
-
-            for text in texts:
-                self.urns.append(text.get("urn"))
-                workUrnsText.append(text.get("workUrn"))
-
-            workUrnsText = [urn for urn in workUrnsText if len(MyCapytain.common.reference.URN(urn)) == 4]
-            self.urns = [urn for urn in self.urns if len(MyCapytain.common.reference.URN(urn)) == 5]
-            self.log("Editions and translations urns : " + " ".join(self.urns))
+        for name, status, op in t.test():
             
-            yield len(worksUrns) == 2 and (len(texts)*2)==len(self.urns + workUrnsText)
-        else:
-            yield False
+            if status:
+                status_str = " passed"
+            else:
+                status_str = " failed"
 
-    def test(self):
-        """ Test a file with various checks
+            logs.append(">>>>> " +name + status_str)
 
-        :returns: List of urns
-        :rtype: list.<str>
-        
-        """
-        self.urns = []
+            if verbose:
+                logs.append("\n".join([o for o in op]))
 
-        for test in INVUnit.tests:
-            # Show the logs and return the status
-            self.flush()
-            try:
-                for status in getattr(self, test)():
-                    yield (INVUnit.readable[test], status, self.logs)
-                    self.flush()
-            except Exception as E:
-                status = False
-                self.error(E)
-                yield (INVUnit.readable[test], status, self.logs)
+            results[f][name] = status
 
-class CTSUnit(TESTUnit):
-    """ CTS testing object
+        results[f] = cover(results[f])
+        passing[f.replace("/", ".")] = True == results[f]["status"]
+        inventory += t.urns
 
-    :param path: Path to the file
-    :type path: basestring
+    else:
+        t = units.CTSUnit(f)
+        logs.append(">>>> Testing "+ f.split("data")[-1])
+        for name, status, op in t.test(tei, epidoc, inventory):
+            
+            if status:
+                status_str = " passed"
+            else:
+                status_str = " failed"
 
+            logs.append(">>>>> " +name + status_str)
+
+            if verbose:
+                logs.append("\n".join([o for o in op]))
+
+            results[f][name] = status
+
+        results[f] = cover(results[f])
+        passing[f.split("/")[-1]] = True == results[f]["status"]
+
+    return logs + ["test+=1"], inventory, results, passing
+
+def run(opts, uuid, reponame, branch, dest):
+    """ Run the tests
+
+    :param opts: Options (-vte)
+    :type opts: str
+    :param uuid: Identifier of the test
+    :type uuid: str
+    :param reponame: Name of the Git Repository
+    :type reponame: str
+    :param branch: Identifier of the branch
+    :type branch: str
+    :param dest: Folder containing the repository
+    :type dest: str
+
+    :returns: Boolean indicating success
+    :rtype: boolean
     """
-    tests = ["parsable", "capitain", "has_urn", "naming_convention", "refsDecl", "passages", "inventory"]
-    readable = {
-        "parsable" : "File parsing",
-        "capitain" : "File ingesting in MyCapytain",
-        "refsDecl" : "RefsDecl parsing",
-        "passages" : "Passage level parsing",
-        "epidoc" : "Epidoc DTD validation",
-        "tei" : "TEI DTD Validation",
-        "has_urn" : "URN informations",
-        "naming_convention" : "Naming conventions",
-        "inventory" : "Available in inventory"
-    }
-    def capitain(self):
-        """ Load the file in MyCapytain
-        """
-        self.Text = MyCapytain.resources.texts.local.Text(resource=self.xml.getroot())
-        if self.Text:
-            yield True
-        else:
-            yield False
+    ###
+    ###    Initialization and parameters recovering
+    ###
+    errors = False
+    inv = []
 
-    def refsDecl(self):
-        """ Contains refsDecl informations
-        """
-        self.log(str(len(self.Text.citation)) + " citations found")
-        yield len(self.Text.citation) > 0
+    PNRT_SETTINGS["directory"] = "/".join([dest, uuid])
+    PNRT_SETTINGS["reponame"] = reponame
 
-    def epidoc(self):
-        test = subprocess.Popen(
-            ["java", "-jar", JING, EPIDOC, self.path],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            shell=False
-        )
+    if len(opts) > 1:
+        opts = list(opts[1:])
+    else:   
+        opts = list()
 
-        out, error = test.communicate()
+    verbose = "v" in opts
+    tei = "t" in opts
+    epidoc = "e" in opts
 
-        if len(out) > 0:
-            for error in out.decode("utf-8").split("\n"):
-                self.log(error)
-        yield len(out) == 0 and len(error) == 0
+    """ 
+        Results storing variables initialization
+    """
 
-    def tei(self):
-        test = subprocess.Popen(
-            ["java", "-jar", JING, TEI_ALL, self.path],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            shell=False
-        )
+    # Store the results
+    results = defaultdict(dict)
+    passing = defaultdict(dict)
 
-        out, error = test.communicate()
+    files, cts__ = utils.find_files(PNRT_SETTINGS["directory"])
 
-        if len(out) > 0:
-            for error in out.decode("utf-8").split("\n"):
-                self.log(error)
-        yield len(out) == 0 and len(error) == 0
+    prnt(">>> Starting tests !")
+    prnt("files="+str(len(files) + len(cts__)))
 
-    def passages(self):
-        for i in range(0, len(self.Text.citation)):
-            passages = self.Text.getValidReff(level=i+1)
-            status = len(passages) > 0
-            self.log(str(len(passages)) + " found")
-            yield status
+    # We load a thread pool which has 5 maximum workers
+    with concurrent.futures.ThreadPoolExecutor(max_workers=5) as executor:
+        # We create a dictionary of tasks which 
+        tasks = {executor.submit(do_test, target_file, tei, epidoc, verbose, inv, results, passing): target_file for target_file in cts__}
+        # We iterate over a dictionary of completed tasks
+        for future in concurrent.futures.as_completed(tasks):
+            logs, inv, results, passing = future.result()
+            for log in logs:
+                prnt(log)
 
-    def has_urn(self):
-        """ Test that a file has its urn saved
-        """
-        urns = self.xml.xpath("//tei:text[starts-with(@n, 'urn:cts:')]", namespaces=ns)
-        urns += self.xml.xpath("//tei:div[starts-with(@n, 'urn:cts:')]", namespaces=ns)
-        status = len(urns) > 0
-        if status:
-            logs = urns[0].get("n")
-            self.log(logs)
-            self.urn = logs
-        yield status
+    # We load a thread pool which has 5 maximum workers
+    with concurrent.futures.ThreadPoolExecutor(max_workers=5) as executor:
+        # We create a dictionary of tasks which 
+        tasks = {executor.submit(do_test, target_file, tei, epidoc, verbose, inv, results, passing): target_file for target_file in files}
+        # We iterate over a dictionary of completed tasks
+        for future in concurrent.futures.as_completed(tasks):
+            logs, inv, results, passing = future.result()
+            for log in logs:
+                prnt(log)
 
-    def naming_convention(self):
-        """ Check the naming convention of the file
-        """
-        if self.urn:
-            yield self.urn.split(":")[-1] in self.path
-        else:
-            yield False
+    prnt(">>> Finished tests !")
 
-    def inventory(self):
-        """ Check the naming convention of the file
-        """
-        if self.urn:
-            yield self.urn in self.inv
-        else:
-            yield False
+    success = len([True for status in passing.values() if status is True])
+    if success == len(passing):
+        status_string = "success"
+    else:
+        status_string = "failure"
+    prnt("[{2}] {0} over {1} texts have fully passed the tests".format(success, len(passing), status_string))
 
-    def test(self, tei, epidoc, inv=None):
-        """ Test a file with various checks
+    prnt("====JSON====")
+    prnt(json.dumps({
+        "status" : success == len(passing),
+        "units" : results,
+        "coverage" : statistics.mean([test["coverage"] for test in results.values()])
+    }))
 
-        :param tei: Test with TEI DTD
-        :type tei: bool
-        :param epidoc: Test with EPIDOC DTD
-        :type epidoc: bool
-        
-        """
-        if inv is None:
-            inv = []
-        self.inv = inv
+    return success == len(passing)
 
-        tests = CTSUnit.tests
-        if tei:
-            tests.append("tei")
-        if epidoc:
-            tests.append("epidoc")
-
-        for test in CTSUnit.tests:
-            # Show the logs and return the status
-            self.flush()
-            try:
-                for status in getattr(self, test)():
-                    yield (CTSUnit.readable[test], status, self.logs)
-                    self.flush()
-            except Exception as E:
-                status = False
-                self.error(E)
-                yield (CTSUnit.readable[test], status, self.logs)
+if __name__ == '__main__':
+    if run(*argv[1:]) is False:
+        exit(1)
+    else:
+        exit(0)
